+++ enclone_visual/src/process_messages.rs
@@ -3,16 +3,15 @@
 use crate::copy_image_to_clipboard::copy_bytes_to_clipboard;
 use crate::history::*;
 use crate::messages::*;
+use crate::proc1::*;
 use crate::share::*;
 use crate::testsuite::TESTS;
 use crate::*;
 use chrono::prelude::*;
-use enclone_core::combine_group_pics::*;
 use flate2::read::GzDecoder;
 use gui_structures::ComputeState::*;
 use iced::{Color, Command};
 use io_utils::*;
-use itertools::Itertools;
 use std::fs::File;
 use std::io::Read;
 use std::time::{Duration, Instant};
@@ -308,141 +307,7 @@ impl EncloneVisual {
             }
 
             Message::SubmitButtonPressed(_) => {
-                let mut group_spec = true;
-                let mut group_ids = Vec::<usize>::new();
-                let s = self.input_value.split(',').collect::<Vec<&str>>();
-                for i in 0..s.len() {
-                    let mut ok = false;
-                    if s[i].parse::<usize>().is_ok() {
-                        let n = s[i].force_usize();
-                        if n >= 1 {
-                            group_ids.push(n);
-                            ok = true;
-                        }
-                    } else if s[i].contains("-") {
-                        let (a, b) = (s[i].before("-"), s[i].after("-"));
-                        if a.parse::<usize>().is_ok() && b.parse::<usize>().is_ok() {
-                            let (a, b) = (a.force_usize(), b.force_usize());
-                            if 1 <= a && a <= b {
-                                for j in a..=b {
-                                    group_ids.push(j);
-                                }
-                                ok = true;
-                            }
-                        }
-                    }
-                    if !ok {
-                        group_spec = false;
-                    }
-                }
-                if group_ids.is_empty() {
-                    group_spec = false;
-                }
-                unique_sort(&mut group_ids);
-                if self.compute_state != WaitingForRequest {
-                    Command::none()
-                } else {
-                    if group_spec {
-                        self.translated_input_value = self.input_value.clone();
-                        let mut reply_text;
-                        let new = self.translated_input_current();
-                        let args = new.split(' ').collect::<Vec<&str>>();
-                        if self.h.input1_history.is_empty() && self.h.input2_history.is_empty() {
-                            reply_text = "Group identifier can only be supplied if another \
-                                command has already been run."
-                                .to_string();
-                        } else if group_ids[group_ids.len() - 1] > self.current_tables.len() {
-                            reply_text = "Group identifier is too large.".to_string();
-                        } else {
-                            let mut group_pics = Vec::<String>::new();
-                            let mut last_widths = Vec::<u32>::new();
-                            for x in group_ids.iter() {
-                                group_pics.push(self.current_tables[*x - 1].clone());
-                                last_widths.push(self.last_widths_value[*x - 1]);
-                            }
-                            reply_text = combine_group_pics(
-                                &group_pics,
-                                &last_widths,
-                                args.contains(&"NOPRINT"),
-                                true, // .noprintx
-                                args.contains(&"HTML"),
-                                args.contains(&"NGROUP"),
-                                false, // .pretty
-                            );
-                            reply_text += "\n \n \n"; // papering over truncation bug in display
-                        }
-                        let mut args2 = Vec::<String>::new();
-                        for x in args.iter() {
-                            if x.len() > 0 && !x.starts_with("G=") {
-                                args2.push(x.to_string());
-                            }
-                        }
-                        args2.push(format!("G={}", self.translated_input_value));
-                        self.output_value = reply_text.to_string();
-                        let hi = self.h.history_index;
-                        self.h
-                            .input1_history
-                            .insert(hi as usize, self.h.input1_hist_uniq.len() as u32);
-                        self.h.input1_hist_uniq.push(self.input1_value.clone());
-                        self.h
-                            .input2_history
-                            .insert(hi as usize, self.h.input2_hist_uniq.len() as u32);
-                        self.h.input2_hist_uniq.push(self.input2_value.clone());
-                        self.h
-                            .translated_input_history
-                            .insert(hi as usize, self.h.translated_input_hist_uniq.len() as u32);
-                        self.h
-                            .translated_input_hist_uniq
-                            .push(args2.iter().format(" ").to_string());
-                        self.h
-                            .svg_history
-                            .insert(hi as usize, self.h.svg_history[(hi - 1) as usize]);
-                        self.h
-                            .summary_history
-                            .insert(hi as usize, self.h.summary_history[(hi - 1) as usize]);
-                        self.h
-                            .displayed_tables_history
-                            .insert(hi as usize, self.h.displayed_tables_hist_uniq.len() as u32);
-                        self.h
-                            .displayed_tables_hist_uniq
-                            .push(reply_text.to_string());
-                        self.h
-                            .table_comp_history
-                            .insert(hi as usize, self.h.table_comp_history[(hi - 1) as usize]);
-                        self.h
-                            .last_widths_history
-                            .insert(hi as usize, self.h.last_widths_history[(hi - 1) as usize]);
-                        self.h.is_blank.insert(hi as usize, self.is_blank_current());
-                        self.h
-                            .descrip_history
-                            .insert(hi as usize, self.h.descrip_history[(hi - 1) as usize]);
-                        self.h.history_index += 1;
-                        if !TEST_MODE.load(SeqCst) {
-                            Command::none()
-                        } else {
-                            Command::perform(noop0(), Message::Capture)
-                        }
-                    } else {
-                        self.compute_state = Thinking;
-                        self.start_command = Some(Instant::now());
-                        // The following sleep is needed for button text to consistenly update.
-                        thread::sleep(Duration::from_millis(20));
-                        if self.input_value.starts_with('#')
-                            && self.cookbook.contains_key(&self.input_value)
-                        {
-                            self.translated_input_value = self.cookbook[&self.input_value].clone();
-                        } else {
-                            self.translated_input_value = self.input_value.clone();
-                        }
-                        USER_REQUEST.lock().unwrap().clear();
-                        USER_REQUEST
-                            .lock()
-                            .unwrap()
-                            .push(self.translated_input_value.clone());
-                        PROCESSING_REQUEST.store(true, SeqCst);
-                        Command::perform(compute(), Message::ComputationDone)
-                    }
-                }
+                submit_button_pressed(self)
             }
 
             Message::DelButtonPressed(_) => {
